Product の　インスタンス生成のロジックをモデルに移行した
